+++ src/guilds.h
@@ -49,8 +49,8 @@ private:
 			std::string rank;
 			std::string nick;
 		};
-		unsigned long gid;
-		static unsigned long counter;
+		uint32_t gid;
+		static uint32_t counter;
 		std::string gname;
 		std::vector<Member> members;
